feat : 투두 필터링 기능 구현 - todo-reducer를 사용하면 기존의 todos 배열이 변경되기 때문에 useState로 filter의 상태를 변경
