+++ src/components/Button.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function Button({ name, onClick }) {
+  return (
+    <button onClick={onClick}>{name}</button>
+  );
+}
