+++ src/components/Navbar.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function Navbar({ children }) {
+  return (
+    <header>
+      <nav>{children}</nav>
+    </header>
+  );
+}
+
